Refactor scraping code: (1) scrape.js -- utility functions for scrapers; (2) scraper.js -- abstract interface for scrapers; (2) phantom-scrape.js -- PhantomJS scraper implementation.
